+++ test/cmdlineTests/combined_json_with_devdoc/input.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: GPL-3.0
+pragma solidity >=0.0;
+
+/// @dev This is devdoc.
+contract C {}
+
+/// And this is a notice.
+contract D {}