+++ test/cmdlineTests/combined_json_with_userdoc/input.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: GPL-3.0
+pragma solidity >=0.0;
+
+/// @notice Description for users.
+contract C {}
+
+/// @dev Description for developers.
+contract D {}